refactor components and storybooks up to mumbles
