+++ src/Components/LandingPage/LandingPage.stories.tsx
@@ -3,13 +3,10 @@ import React from 'react';
 import { LandingPage } from './LandingPage';
 
 export default {
-  title: 'LandingPage',
+  title: 'Composition/LandingPage',
   component: LandingPage,
 } as ComponentMeta<typeof LandingPage>;
 
 const Template: ComponentStory<typeof LandingPage> = (args) => <LandingPage {...args} />;
 
 export const Landing = Template.bind({});
-Landing.args = {
-  isRegistered: true,
-};
