+++ src/Components/ProfilePicture/ProfilePicture.tsx
@@ -1,18 +1,6 @@
 import React, { FC, useCallback } from 'react';
-
+import { IconColor, ProfilePictureProps } from '../../models';
 import { Icon } from '../Icon';
-import { IconColor } from '../../models';
-
-export type AvatarSize = 'small' | 'medium' | 'large';
-
-interface ProfilePictureProps {
-  name: string;
-  src?: string;
-  size: AvatarSize;
-  edit?: boolean;
-  onClick?: () => void;
-  onEdit?: () => void;
-}
 
 export const ProfilePicture: FC<ProfilePictureProps> = ({ name, src = '', size, edit = false, onClick, onEdit }) => {
   const getClasses = useCallback((size: string) => {
