+++ src/Components/UserWidget/UserWidget.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { UserWidget } from './UserWidget';
 
 export default {
-  title: 'UserWidget',
+  title: 'Composition/UserWidget',
   component: UserWidget,
 } as ComponentMeta<typeof UserWidget>;
 
@@ -12,18 +12,13 @@ const Template: ComponentStory<typeof UserWidget> = (args) => <UserWidget {...ar
 const TemplateRow: ComponentStory<typeof UserWidget> = (args) => (
   <div className="flex flex-row items-center justify-around">
     <div className="p-2 w-56">
-      <UserWidget src="assets/images/profile/r.vogt.jpg" name="Damian Caduff" size="small" username="damiancaduff" />
+      <UserWidget src="assets/images/profile/r.vogt.jpg" name="Damian Caduff" username="damiancaduff" />
     </div>
     <div className="p-2 w-56">
-      <UserWidget src="assets/images/profile/r.vogt.jpg" name="Robert Vogt" size="small" username="robertvogt" />
+      <UserWidget src="assets/images/profile/r.vogt.jpg" name="Robert Vogt" username="robertvogt" />
     </div>
     <div className="p-2 w-56">
-      <UserWidget
-        src="assets/images/profile/r.vogt.jpg"
-        name="Chrisoph Bühler"
-        size="small"
-        username="christophbuehler"
-      />
+      <UserWidget src="assets/images/profile/r.vogt.jpg" name="Chrisoph Bühler" username="christophbuehler" />
     </div>
   </div>
 );
@@ -32,7 +27,6 @@ export const Widget = Template.bind({});
 Widget.args = {
   src: 'assets/images/profile/r.vogt.jpg',
   name: 'Robert Vogt',
-  size: 'small',
   username: 'robertvogt',
 };
