+++ src/Components/UserWidget/UserWidget.tsx
@@ -1,33 +1,17 @@
 import React, { FC } from 'react';
 import { v4 as uuid } from 'uuid';
-
-import { IconColor, IconHoverColor, TextColor, TextHoverColor } from '../../models';
+import { IconColor, IconHoverColor, TextColor, TextHoverColor, UserWidgetProps } from '../../models';
 import { Button } from '../Button';
 import { Interaction } from '../Interaction';
-import { AvatarSize, ProfilePicture } from '../ProfilePicture';
+import { ProfilePicture } from '../ProfilePicture';
 
-interface UserWidgetProps {
-  name: string;
-  username: string;
-  src: string;
-  size: AvatarSize;
-  edit?: boolean;
-  onClick?: () => void;
-  onEdit?: () => void;
-}
-const defaultProps: Partial<UserWidgetProps> = {
-  edit: false,
-};
-
-const userImage = 'assets/images/profile/r.vogt.jpg';
-
-export const UserWidget: FC<UserWidgetProps> = ({ name, username, src, size, edit, onClick, onEdit }) => {
+export const UserWidget: FC<UserWidgetProps> = ({ name, username, src, onClick }) => {
   const inputId = uuid();
 
   return (
     <div className="bg-white rounded-2xl relative p-4 w-full">
       <div className="p-2 place-content-center flex">
-        <ProfilePicture name="rvogt" size="medium" src={userImage} />
+        <ProfilePicture name="rvogt" size="medium" src={src} />
       </div>
       <h4
         className="p-1 text-xl text-center  whitespace-nowrap text-ellipsis overflow-hidden font-semibold"
@@ -52,12 +36,10 @@ export const UserWidget: FC<UserWidgetProps> = ({ name, username, src, size, edi
           color="Violet"
           size="M"
           label="Follow"
-          onClick={() => console.log('follow')}
+          onClick={onClick}
           icon="mumble"
         />
       </div>
     </div>
   );
 };
-
-UserWidget.defaultProps = defaultProps;
